refactor(chrome): prepare host-aware presentation across the application

+++ src/ui/appearance_exerciser_tests.rs
@@ -211,7 +211,7 @@ fn replace_appearance(generation: u64, cx: &mut VisualTestContext) {
             spaceterm_ui::replace_control_theme_catalog(cx, controls),
             Ok(spaceterm_ui::ControlThemeReplacement::Applied)
         );
-        cx.set_global(InstalledChrome(Arc::new(appearance)));
+        cx.set_global(InstalledChrome::single(Arc::new(appearance)));
         window.refresh();
     });
     cx.run_until_parked();
